+++ service/controller/handler/configversion/delete.go
@@ -1,0 +1,9 @@
+package configversion
+
+import (
+	"context"
+)
+
+func (h *Handler) EnsureDeleted(ctx context.Context, obj interface{}) error {
+	return nil
+}