+++ service/controller/handler/configversion/types.go
@@ -1,0 +1,10 @@
+package configversion
+
+type Index struct {
+	Entries map[string][]IndexEntry `json:"entries"`
+}
+
+type IndexEntry struct {
+	ConfigVersion string `json:"configVersion,omitempty"`
+	Version       string `json:"version"`
+}