+++ service/controller/handler/values/create.go
@@ -9,34 +9,47 @@ import (
 	"github.com/giantswarm/microerror"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 
-	controllerkey "github.com/giantswarm/config-controller/service/controller/key"
+	"github.com/giantswarm/config-controller/service/controller/key"
 )
 
 func (h *Handler) EnsureCreated(ctx context.Context, obj interface{}) error {
-	app, err := controllerkey.ToAppCR(obj)
+	app, err := key.ToAppCR(obj)
 	if err != nil {
 		return microerror.Mask(err)
 	}
+	annotations := app.GetAnnotations()
+	if annotations == nil {
+		annotations = map[string]string{}
+	}
 
-	configVersion, ok := app.GetAnnotations()[annotation.ConfigVersion]
+	configVersion, ok := annotations[annotation.ConfigVersion]
 	if !ok {
-		h.logger.Debugf(ctx, "App CR %q is missing %q annotation", app.Name, annotation.ConfigVersion)
-		if _, ok := app.GetAnnotations()[PauseAnnotation]; ok {
-			err = h.removeAnnotation(ctx, &app, PauseAnnotation)
+		h.logger.Debugf(ctx, "App CR is missing %q annotation", annotation.ConfigVersion)
+		h.logger.Debugf(ctx, "cancelling handler")
+		return nil
+	}
+
+	if configVersion == key.LegacyConfigVersion {
+		h.logger.Debugf(ctx, "App CR has config version %#q", configVersion)
+		if _, ok := annotations[annotation.AppOperatorPaused]; ok {
+			h.logger.Debugf(ctx, "App does not use generated config, removing pause annotation")
+			app.SetAnnotations(key.RemoveAnnotation(annotations, annotation.AppOperatorPaused))
+			err = h.k8sClient.CtrlClient().Update(ctx, &app)
 			if err != nil {
-				return err
+				return microerror.Mask(err)
 			}
+			h.logger.Debugf(ctx, "removed %#q annotation", annotation.AppOperatorPaused)
 		}
 		h.logger.Debugf(ctx, "cancelling handler")
 		return nil
 	}
 
-	h.logger.Debugf(ctx, "generating app %#q config version %#q", app.Spec.Name, configVersion)
+	h.logger.Debugf(ctx, "generating app config version %#q", configVersion)
 	configmap, secret, err := h.generateConfig(ctx, h.installation, app.Namespace, app.Spec.Name, configVersion)
 	if err != nil {
 		return microerror.Mask(err)
 	}
-	h.logger.Debugf(ctx, "generated app %#q config version %#q", app.Spec.Name, configVersion)
+	h.logger.Debugf(ctx, "generated app config version %#q", configVersion)
 
 	h.logger.Debugf(ctx, "ensuring configmap %s/%s", configmap.Namespace, configmap.Name)
 	err = h.k8sClient.CtrlClient().Create(ctx, configmap)
@@ -71,32 +84,16 @@ func (h *Handler) EnsureCreated(ctx context.Context, obj interface{}) error {
 		Name:      secret.Name,
 	}
 	if !reflect.DeepEqual(app.Spec.Config.ConfigMap, configmapReference) || !reflect.DeepEqual(app.Spec.Config.Secret, secretReference) {
-		h.logger.Debugf(ctx, "updating App CR %#q with configmap and secret details", app.Name)
-		app.SetAnnotations(removeAnnotation(app.GetAnnotations(), PauseAnnotation))
+		h.logger.Debugf(ctx, "updating App CR with configmap and secret details")
+		app.SetAnnotations(key.RemoveAnnotation(annotations, annotation.AppOperatorPaused))
 		app.Spec.Config.ConfigMap = configmapReference
 		app.Spec.Config.Secret = secretReference
 		err = h.k8sClient.CtrlClient().Update(ctx, &app)
 		if err != nil {
 			return microerror.Mask(err)
 		}
-		h.logger.Debugf(ctx, "updated App CR %#q with configmap and secret details", app.Name)
+		h.logger.Debugf(ctx, "updated App CR with configmap and secret details")
 	}
 
 	return nil
-}
-
-func removeAnnotation(annotations map[string]string, key string) map[string]string {
-	if annotations == nil {
-		return nil
-	}
-
-	out := map[string]string{}
-	for k, v := range annotations {
-		if k == key {
-			continue
-		}
-		out[k] = v
-	}
-
-	return out
 }