+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Allow caching tags and pulled repositories.
 - Handle `app-operator.giantswarm.io/pause` annotation.
 - Clear `app-operator.giantswarm.io/pause` if App CR does is not annotated with config version.
+- Annotate App CR with config version defined in catalog's `index.yaml`.
 
 ## [0.1.0] - 2020-11-26